# app/backend/src/types/user.types.ts
export const USER_ROLES = ['customer', 'admin'] as const;

export type UserRole = (typeof USER_ROLES)[number];

export interface PublicUser {
  id: number;
  firstName: string;
  lastName: string;
  username: string;
  role: UserRole;
}

export interface StoredUser extends PublicUser {
  passwordHash: string;
  passwordVersion: number;
}

export interface NewUser {
  firstName: string;
  lastName: string;
  username: string;
  password: string;
  role?: UserRole | undefined;
}

export interface NewUserRow extends Omit<NewUser, 'password'> {
  passwordHash: string;
  passwordVersion: number;
}

export interface ProfileUpdate {
  firstName?: string | undefined;
  lastName?: string | undefined;
  username?: string | undefined;
}